--- features/auth/presentation/src/main/java/com/fredprojects/features/auth/presentation/vm/AuthEvents.kt
@@ -1,9 +1,0 @@
-package com.fredprojects.features.auth.presentation.vm
-
-import com.fredprojects.features.auth.presentation.models.UDPModel
-
-sealed class AuthEvents {
-    data class Authorization(val login: String, val password: String) : AuthEvents()
-    data class UpsertUserData(val user: UDPModel) : AuthEvents()
-    data object DeleteUser : AuthEvents()
-}
