+++ features/auth/src/main/AndroidManifest.xml
