+++ features/auth/src/main/java/com/fredprojects/features/auth/domain/utils/AuthStatus.kt
